+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 All Notable changes to `drupal/opening-hours` module.
 
-## [Unreleased]
+## [8.x-1.0-aplha1]
 
 ### Added
 * DMOH-19: Added service configuration form.
+* DMOH-20: Added the opening hours field type.
+* DMOH-21: Added the opening hours field widget.
 
+[8.x-1.0-aplha1]: https://github.com/StadGent/drupal_module_opening-hours/releases/tag/8.x-1.0-aplha1
 [Unreleased]: https://github.com/StadGent/drupal_module_opening-hours/compare/master...develop